% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_validation.R
\name{summarize_adaptme_validation}
\alias{summarize_adaptme_validation}
\title{Aggregate ADAPT-ME Validation Metrics}
\usage{
summarize_adaptme_validation(validation_results)
}
\arguments{
\item{validation_results}{output from \code{run_adaptme_simulation_validation}}
}
\value{
A scenario-level summary data frame.
}
\description{
Aggregate ADAPT-ME Validation Metrics
}
